Support global conf, kafka ssl
